Refactor/import file (#24)
* chore: implicit conversion build error

```
simulate/hook_draw.c:73:40: error: implicit conversion from 'int' to 'float' changes value from 99999999 to 1.0E+8 [-Werror,-Wimplicit-const-int-float-conversion]
        if (hit(world, ray, (t_range){ DELTA, BIGVALUE }, &rec)) {
                                     ~        ^~~~~~~~
simulate/hook_draw.c:15:18: note: expanded from macro 'BIGVALUE'
                 ^~~~~~~~
1 error generated.
```

* refactor: Move debug functions

* refactor: move parse function to file directory

* refactor: about struct s_world

* refactor: Wimplicit-const-int-float-conversion

* refactor: delete float to float casting

---------
