+++ src/Element/AmbientLight/ambient_light.c
@@ -6,18 +6,29 @@
 /*   By: yonshin <yonshin@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 12:04:44 by yonshin           #+#    #+#             */
-/*   Updated: 2023/06/10 01:40:09 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:40 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ambient_light.h"
+
+int	debug_ambient_light(void *param)
+{
+	struct s_ambient_light *const	c = param;
+
+	if (c == NULL || DEBUG == 0)
+		return (EXIT_SUCCESS);
+	printf("%s: ratio[%f]\n", __func__, c->raw.ratio);
+	printf("%s: rgb[%d, %d, %d]\n", __func__,
+		c->raw.rgb.r, c->raw.rgb.g, c->raw.rgb.b);
+	return (EXIT_SUCCESS);
+}
 
 t_func	ambient_light(enum e_element_func func)
 {
 	static const t_func	functions[MAX_ELEMENT_TYPE] = {
 	[Init] = init_ambient_light,
 	[Hit] = hit_ambient_light,
-	[Parse] = parse_ambient_light,
 	[Debug] = debug_ambient_light};
 
 	return (functions[func]);
