+++ src/Element/Light/light.h
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 05:37:19 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 01:51:23 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:08 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,6 @@ typedef struct s_light
 
 extern t_func	light(enum e_element_func func);
 extern int		debug_light(void *param);
-extern int		parse_light(void *param, char **argv);
 extern int		hit_light(void *this, t_ray *r, t_range ra, t_hit *h);
 extern int		init_light(t_light *self);
