+++ src/Element/Sphere/sphere.h
@@ -6,7 +6,7 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/31 05:37:44 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 01:51:44 by minjungk         ###   ########.fr       */
+/*   Updated: 2023/07/03 22:43:14 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,6 @@ typedef struct s_sphere
 
 extern t_func	sphere(enum e_element_func func);
 extern int		debug_sphere(void *param);
-extern int		parse_sphere(void *param, char **argv);
 extern int		hit_sphere(void *this, t_ray *r, t_range ra, t_hit *h);
 extern int		init_sphere(t_sphere *self);
