+++ src/file/parse_light.c
@@ -1,46 +1,34 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   sphere.c                                           :+:      :+:    :+:   */
+/*   parse_light.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: minjungk <minjungk@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/31 05:37:44 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 00:49:24 by minjungk         ###   ########.fr       */
+/*   Created: 2023/05/31 05:37:19 by minjungk          #+#    #+#             */
+/*   Updated: 2023/07/04 05:32:53 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "sphere.h"
+#include "../Element/Light/light.h"
 
-int	debug_sphere(void *param)
-{
-	struct s_sphere *const	c = param;
-
-	if (c == NULL || DEBUG == 0)
-		return (EXIT_SUCCESS);
-	printf("%s: coordinate[%f, %f, %f]\n", __func__,
-		c->raw.coordinate.x, c->raw.coordinate.y, c->raw.coordinate.z);
-	printf("%s: diameter[%f]\n", __func__, c->raw.diameter);
-	printf("%s: rgb[%d, %d, %d]\n", __func__,
-		c->raw.rgb.r, c->raw.rgb.g, c->raw.rgb.b);
-	return (EXIT_SUCCESS);
-}
-
-int	parse_sphere(void *param, char **argv)
+int	parse_light(void *param, char **argv)
 {
 	char					*remain;
-	struct s_sphere *const	content = param;
+	struct s_light *const	content = param;
 
-	if (!argv || !argv[0] || !argv[1] || !argv[2])
+	if (!argv || !argv[0] || !argv[1])
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	content->type = Sphere;
+	content->type = Light;
 	if (parse_vector3(&content->raw.coordinate, argv[0], AllScope))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	content->raw.diameter = ft_strtof(argv[1], &remain);
+	content->raw.ratio = ft_strtof(argv[1], &remain);
 	if (errno || (*remain != '\0' && ft_strchr("\r\n", *remain) == NULL))
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	if (parse_rgb(&content->raw.rgb, argv[2]))
+	if (!(0.0 <= content->raw.ratio && content->raw.ratio <= 1.0))
+		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
+	if (argv[2] && parse_rgb(&content->raw.rgb, argv[2]))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	debug_sphere(content);
+	debug_light(content);
 	return (EXIT_SUCCESS);
 }
