+++ src/file/parse_plane.c
@@ -1,20 +1,20 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   cylinder.c                                         :+:      :+:    :+:   */
+/*   parse_plane.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: minjungk <minjungk@student.42seoul.>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/05/31 05:38:49 by minjungk          #+#    #+#             */
-/*   Updated: 2023/06/10 00:48:45 by minjungk         ###   ########.fr       */
+/*   Created: 2023/05/31 05:38:22 by minjungk          #+#    #+#             */
+/*   Updated: 2023/07/04 05:32:56 by minjungk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "cylinder.h"
+#include "../Element/Plane/plane.h"
 
-int	debug_cylinder(void *param)
+int	debug_plane(void *param)
 {
-	struct s_cylinder *const	c = param;
+	struct s_plane *const	c = param;
 
 	if (c == NULL || DEBUG == 0)
 		return (EXIT_SUCCESS);
@@ -22,33 +22,24 @@ int	debug_cylinder(void *param)
 		c->raw.coordinate.x, c->raw.coordinate.y, c->raw.coordinate.z);
 	printf("%s: axis[%f, %f, %f]\n", __func__,
 		c->raw.axis.x, c->raw.axis.y, c->raw.axis.z);
-	printf("%s: diameter[%f]\n", __func__, c->raw.diameter);
-	printf("%s: height[%f]\n", __func__, c->raw.height);
 	printf("%s: rgb[%d, %d, %d]\n", __func__,
 		c->raw.rgb.r, c->raw.rgb.g, c->raw.rgb.b);
 	return (EXIT_SUCCESS);
 }
 
-int	parse_cylinder(void *param, char **argv)
+int	parse_plane(void *param, char **argv)
 {
-	char						*remain;
-	struct s_cylinder *const	content = param;
+	struct s_plane *const	content = param;
 
-	if (!argv || !argv[0] || !argv[1] || !argv[2] || !argv[3] || !argv[4])
+	if (!argv || !argv[0] || !argv[1] || !argv[2])
 		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	content->type = Cylinder;
+	content->type = Plane;
 	if (parse_vector3(&content->raw.coordinate, argv[0], AllScope))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
 	if (parse_vector3(&content->raw.axis, argv[1], UnitScope))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	content->raw.diameter = ft_strtof(argv[2], &remain);
-	if (errno || (*remain != '\0' && ft_strchr("\r\n", *remain) == NULL))
-		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	content->raw.height = ft_strtof(argv[3], &remain);
-	if (errno || (*remain != '\0' && ft_strchr("\r\n", *remain) == NULL))
-		return (ft_error(__func__, __FILE__, __LINE__, EINVAL));
-	if (parse_rgb(&content->raw.rgb, argv[4]))
+	if (parse_rgb(&content->raw.rgb, argv[2]))
 		return (ft_error(__func__, __FILE__, __LINE__, 0));
-	debug_cylinder(content);
+	debug_plane(content);
 	return (EXIT_SUCCESS);
 }
